DateTimeUtil: Replace unused date format d-M-yyyy
The d-M-yyyy format (e.g. 30-12-2020) is not used in the application
because certain fields require a hyphen. An example is the PeriodOfStay
field in Person, which takes a date range separated by a hyphen (e.g. 20201220-20201230).
Using the d-M-yyyy results in too many hyphens in these dates and may confuse the user.

Let's replace the d-M-yyyy format with a more commonly used format: d MMM yyyy
(e.g. 30 Dec 2020). Dates following this format are pleasant to read for the user.