ParserUtil: Refactor patient-related methods
ParserUtil contains methods for both general purposes and parsing
patient-specific fields.

This results in a class that is unnecessary large.

Let's refactor all patient-related methods into PatientParserUtil, keeping
only the general methods in ParserUtil. In addition, RoomParserUtil::parseRoomNumber
can be treated as a general method and is refactored into ParserUtil::parsePositiveInteger.
Similarly, the AddPatientCommand::arePrefixesPresent method is moved to
ParserUtil.

After refactoring, the codebase will be more consistent with TaskParserUtil.