Update UI for Task to display room and task numbers
Room and task numbers are not shown in the UI for task. Setting room and
task numbers in the UI has been difficult because a global TaskList
(specifically Model#TaskList) is used for displaying the tasks but this
TaskList has no connection with Room, even though TaskList is an
attribute of Room.

Since the task-related commands (e.g. addtask, deletetask, etc.) rely
on the room and task numbers in order to function, the UI for task is
not very useful and cannot be referenced when a user wants to run these
commands.

Let's use another implementation whereby we avoid the usage of
the global TaskList, and instead retrieve Task details directly from
the global list of Room (RoomList). We can use a ScrollPane and manually
insert the lists of tasks from every Room as a child node. We can also
set a listener to refresh the ScrollPane after every change to Room
or the tasks within.

A ScrollPane is used instead of nested ListView. (i.e. ListView<Room>
with each cell containing ListView<Task>) because the number of tasks
in each Room varies largely. This leads to a problem with the
height of ListView<Room> not being set dynamically. Using a ScrollPane
fixes that issue.

Because the ScrollPane is regenerated after the listener catches a
change, there may be performance issues. It is recommended
that some form of caching be used in the future. Perhaps a HashMap
to store the children of the nodes, and after every change, replace
only the nodes that need to be changed.