Add abstraction layer RoomTasks
Each room has an attribute TaskList representing the list of tasks in
the room.

However, the tasks in a room may have different behaviors
compared to a regular list of tasks. Since these behaviors should
not be defined in TaskList, an abstraction is necessary.

Let's add a class RoomTasks to keep track of the tasks in a room.
RoomTasks will wrap a TaskList and each room now has an attribute
RooMTasks instead of TaskList.