Simplify logic for task-related operations
CRUD for tasks works in the following manner:
1. Find room in RoomList
2. If room is found, run the CRUD operation through Model
3. Model routes the request to RoomList
4. RoomList checks if it contains the room and task
5. If room and task are found, RoomList routes the request to
target room
6. Room performs the CRUD operation and returns control to RoomList
7. RoomList replaces original room with the updated room

The above relies excessively on RoomList and triggers a change
in RoomList every time there is a task operation. This is not
ideal for UI as a listener on RoomList will be notified whenever
there is a task operation.

Let's use the following flow instead:
1. Find room in RoomList
2. If room is found, run the CRUD operation through Model
3. Model checks with RoomList to verify room is in RoomList
4. Model routes the request to room alongside the task
5. Room performs the CRUD operation

The new logic is simpler and does not trigger changes in RoomList
for task operations.