Revert the adding of room numbers to tasks.
Tasks currently have a room number attribute. This attribute is used only
to display the room number in the user interface.

This may lead to synchronization issues if the users attempt to modify
the value of this attribute in the .json file, e.g. a task in room 4
changed to be in room 3. This poses a problem in the user interface as
the label for room number will wrongly display 3 instead of 4 (the actual
room the task is in).

Let's move the room number to an association class to avoid this issue.